+++ carp_mobile_sensing/example/lib/paper_example.dart
@@ -4,7 +4,7 @@ import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
 void sensing() async {
   // create the study
-  Study study = Study("2", 'user@gmail.com',
+  Study study = Study('2', 'user@gmail.com',
       name: 'A simple example study',
       dataEndPoint: FileDataEndPoint()
         ..bufferSize = 500 * 1000