+++ carp_mobile_sensing/lib/runtime/study_controller.dart
@@ -81,8 +81,8 @@ class StudyController {
     // 2. preferred data format as specified in the study protocol
     // 3. any custom transformer
     events = transformer(executor.events
-        .map((datum) => TransformerSchemaRegistry.lookup(privacySchemaName).transform(datum))
-        .map((datum) => TransformerSchemaRegistry.lookup(study.dataFormat).transform(datum)));
+        .map((datum) => TransformerSchemaRegistry.instance.lookup(privacySchemaName).transform(datum))
+        .map((datum) => TransformerSchemaRegistry.instance.lookup(study.dataFormat).transform(datum)));
 
     // old, simple version below
     // events = transformer(executor.events);
@@ -95,8 +95,8 @@ class StudyController {
     Device.getDeviceInfo();
 
     // setting up permissions
-    permissions = await PermissionHandlerPlatform.instance.requestPermissions(SamplingPackageRegistry.permissions);
-    SamplingPackageRegistry.permissions.forEach((permission) {
+    permissions = await PermissionHandlerPlatform.instance.requestPermissions(SamplingPackageRegistry.instance.permissions);
+    SamplingPackageRegistry.instance.permissions.forEach((permission) {
       PermissionStatus status = permissions[permission];
       if (status != PermissionStatus.granted) {
         warning('Permissions not granted for $permission, permission is $status');